refactor(transactions): move logic of transaction out the block indexer
