+++ packages/block-indexer/src/modules/icon-indexer/block-indexer.js
@@ -3,10 +3,10 @@
 const debug = require('debug')('icon');
 const IconService = require('icon-sdk-js');
 const { HttpProvider } = require('icon-sdk-js');
-const { logger, RESULT_CODE } = require('../../common');
-const { hexToDecimal } = require('../../common/util');
-const { saveBlock, saveTransaction, setTransactionConfirmed, getBySerialNumber, getLastBlock } = require('./repository');
+const { logger } = require('../../common');
+const { saveBlock, getLastBlock } = require('./repository');
 const { handleAuctionEvents } = require('./auctions');
+const { handleTransEvent } = require('./transactions');
 
 const httpProvider = new HttpProvider(process.env.ICON_API_URL);
 const iconService = new IconService(httpProvider);
@@ -14,53 +14,6 @@ const iconService = new IconService(httpProvider);
 let blockHeight = Number(process.env.ICON_BLOCK_HEIGHT);
 let isWaitToStop = false;
 
-/**
- * Confirm TransferEnd event
- * @param {*} event
- * @param {*} txInfo
- */
-async function confirmTransfEnd(event, txInfo) {
-  let indexedData = event.indexed;
-  let data = event.data;
-  let transaction = await getBySerialNumber(hexToDecimal(indexedData[1]));
-  // has transaction and transfer result code must be 0
-  if (transaction && hexToDecimal(data[0]) == RESULT_CODE.RC_OK) {
-    await setTransactionConfirmed([transaction], txInfo)
-  }
-}
-
-/**
- * Handle TransferStart and TransferEnd events
- * @param {*} txResult
- */
-async function handleTransEvent(txResult) {
-  let transactions = [];
-
-  for (let event of txResult.eventLogs) {
-    //handle TransactionStart event
-    if (event.indexed.find(item => item.includes('TransferStart'))) {
-      let indexedData = event.indexed;
-      let data = event.data;
-      let transObj = {
-        fromAddress: indexedData[1],
-        tokenName: indexedData[2],
-        serialNumber: hexToDecimal(data[0]),
-        value: hexToDecimal(data[1]),
-        toAddress: data[2],
-        txHash: txResult.txHash,
-        blockHash: txResult.blockHash,
-        blockHeight: txResult.blockHeight,
-        confirmed: false,
-      }
-      transactions.push(transObj)
-    } else if ((event.indexed.find(item => item.includes('TransferEnd')))) {
-      confirmTransfEnd(event, { txHash: txResult.txHash, blockHeight: txResult.blockHeight, blockHash: txResult.blockHash })
-    }
-  }
-  if (transactions.length > 0) {
-    await saveTransaction(transactions);
-  }
-}
 
 async function runTransactionHandlers(transaction, txResult, block) {
   await handleTransEvent(txResult);
